Split cluster tests into separate class.
