+++ src/test/java/com/google/jenkins/plugins/k8sengine/KubernetesEngineBuilderTest.java
@@ -24,13 +24,11 @@ import static org.mockito.ArgumentMatchers.anyString;
 
 import com.google.api.services.cloudresourcemanager.model.Project;
 import com.google.api.services.compute.model.Zone;
-import com.google.api.services.container.model.Cluster;
 import com.google.common.collect.ImmutableList;
 import com.google.jenkins.plugins.k8sengine.KubernetesEngineBuilder.DescriptorImpl;
 import com.google.jenkins.plugins.k8sengine.client.ClientFactory;
 import com.google.jenkins.plugins.k8sengine.client.CloudResourceManagerClient;
 import com.google.jenkins.plugins.k8sengine.client.ComputeClient;
-import com.google.jenkins.plugins.k8sengine.client.ContainerClient;
 import hudson.AbortException;
 import hudson.util.FormValidation;
 import hudson.util.ListBoxModel;
@@ -48,13 +46,11 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
 public class KubernetesEngineBuilderTest {
-  private static final String TEST_ZONE_A = "us-west1-a";
-  private static final String TEST_ZONE_B = "us-central1-b";
-  private static final String TEST_PROJECT_ID = "test-project-id";
-  private static final String OTHER_PROJECT_ID = "other-project-id";
-  private static final String TEST_CLUSTER = "testCluster (us-west1-a)";
-  private static final String OTHER_CLUSTER = "otherCluster (us-central1-b)";
-  private static final String TEST_CREDENTIALS_ID = "test-credentials-id";
+  static final String TEST_ZONE_A = "us-west1-a";
+  static final String TEST_ZONE_B = "us-central1-b";
+  static final String TEST_PROJECT_ID = "test-project-id";
+  static final String OTHER_PROJECT_ID = "other-project-id";
+  static final String TEST_CREDENTIALS_ID = "test-credentials-id";
 
   private static Jenkins jenkins;
 
@@ -442,273 +438,6 @@ public class KubernetesEngineBuilderTest {
   }
 
   @Test
-  public void testDoFillClusterItemsEmptyWithEmptyCredentialsId() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    ListBoxModel expected =
-        initExpected(ImmutableList.of(EMPTY_NAME), ImmutableList.of(EMPTY_VALUE));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(jenkins, null, null, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-  }
-
-  @Test
-  public void testDoFillClusterItemsEmptyWithEmptyProjectId() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    ListBoxModel expected =
-        initExpected(ImmutableList.of(EMPTY_NAME), ImmutableList.of(EMPTY_VALUE));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(jenkins, null, TEST_CREDENTIALS_ID, null, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-  }
-
-  @Test
-  public void testDoFillClusterItemsErrorMessageWithAbortException() throws IOException {
-    DescriptorImpl descriptor =
-        setUpClusterDescriptor(ImmutableList.of(), new AbortException(), null);
-    ListBoxModel expected =
-        initExpected(
-            ImmutableList.of(Messages.KubernetesEngineBuilder_CredentialAuthFailed()),
-            ImmutableList.of(EMPTY_VALUE));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithIOException() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, new IOException());
-    ListBoxModel expected =
-        initExpected(
-            ImmutableList.of(Messages.KubernetesEngineBuilder_ClusterFillError()),
-            ImmutableList.of(EMPTY_VALUE));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithInvalidCluster() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(TEST_CLUSTER), null, null);
-    ListBoxModel expected =
-        initExpected(
-            ImmutableList.of(EMPTY_NAME, TEST_CLUSTER),
-            ImmutableList.of(EMPTY_VALUE, TEST_CLUSTER));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, "wrong", TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-    assertValueSelected(result, TEST_CLUSTER);
-  }
-
-  @Test
-  public void testDoFillClusterItemsEmptyWithValidInputsNoClusters() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    ListBoxModel expected =
-        initExpected(ImmutableList.of(EMPTY_NAME), ImmutableList.of(EMPTY_VALUE));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, null, TEST_CREDENTIALS_ID, OTHER_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithValidInputsOneCluster() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(TEST_CLUSTER), null, null);
-    ListBoxModel expected =
-        initExpected(
-            ImmutableList.of(EMPTY_NAME, TEST_CLUSTER),
-            ImmutableList.of(EMPTY_VALUE, TEST_CLUSTER));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-    assertValueSelected(result, TEST_CLUSTER);
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithValidInputsMultipleClusters() throws IOException {
-    DescriptorImpl descriptor =
-        setUpClusterDescriptor(ImmutableList.of(OTHER_CLUSTER, TEST_CLUSTER), null, null);
-    ListBoxModel expected =
-        initExpected(
-            ImmutableList.of(EMPTY_NAME, OTHER_CLUSTER, TEST_CLUSTER),
-            ImmutableList.of(EMPTY_VALUE, OTHER_CLUSTER, TEST_CLUSTER));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-    assertValueSelected(result, OTHER_CLUSTER);
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithValidInputsMultipleClustersAndPreviousValue()
-      throws IOException {
-    DescriptorImpl descriptor =
-        setUpClusterDescriptor(ImmutableList.of(OTHER_CLUSTER, TEST_CLUSTER), null, null);
-    ListBoxModel expected =
-        initExpected(
-            ImmutableList.of(EMPTY_NAME, OTHER_CLUSTER, TEST_CLUSTER),
-            ImmutableList.of(EMPTY_VALUE, OTHER_CLUSTER, TEST_CLUSTER));
-    ListBoxModel result =
-        descriptor.doFillClusterItems(
-            jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertListBoxModelEquals(expected, result);
-    assertValueSelected(result, TEST_CLUSTER);
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithNonEmptyZone() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(TEST_CLUSTER), null, null);
-    {
-      ListBoxModel expected =
-          initExpected(
-              ImmutableList.of(EMPTY_NAME, TEST_CLUSTER),
-              ImmutableList.of(EMPTY_VALUE, TEST_CLUSTER));
-      ListBoxModel result =
-          descriptor.doFillClusterItems(
-              jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, TEST_ZONE_A);
-      assertNotNull(result);
-      assertListBoxModelEquals(expected, result);
-      assertValueSelected(result, TEST_CLUSTER);
-    }
-  }
-
-  @Test
-  public void testDoFillClusterItemsWithNonEmptyZoneAndPreviousValue() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    {
-      ListBoxModel expected =
-          initExpected(ImmutableList.of(EMPTY_NAME), ImmutableList.of(EMPTY_VALUE));
-      // The scenario is when the user had previously selected cluster then changed zone
-      ListBoxModel result =
-          descriptor.doFillClusterItems(
-              jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, TEST_ZONE_B);
-      assertNotNull(result);
-      assertListBoxModelEquals(expected, result);
-    }
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithEmptyCluster() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_ClusterRequired(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithEmptyCredentialsId() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(jenkins, TEST_CLUSTER, null, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(
-        Messages.KubernetesEngineBuilder_ClusterCredentialIDRequired(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithEmptyProjectId() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, null, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_ClusterProjectIDRequired(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithValidInputsNoClusters() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(
-            jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_NoClusterInProjectZone(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterOkWithValidInputs() throws IOException {
-    DescriptorImpl descriptor =
-        setUpClusterDescriptor(ImmutableList.of(OTHER_CLUSTER, TEST_CLUSTER), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(
-            jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(FormValidation.ok(), result);
-  }
-
-  @Test
-  public void testDoCheckClusterOkWithNonEmptyMatchingZone() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(OTHER_CLUSTER), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(
-            jenkins, OTHER_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, TEST_ZONE_B);
-    assertNotNull(result);
-    assertEquals(FormValidation.ok(), result);
-  }
-
-  @Test
-  public void testDoCHeckClusterMessageWithNonEmptyNonMatchingZone() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(TEST_CLUSTER), null, null);
-    FormValidation result =
-        descriptor.doCheckCluster(
-            jenkins, OTHER_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, TEST_ZONE_A);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_ClusterNotInProjectZone(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithAbortException() throws IOException {
-    DescriptorImpl descriptor =
-        setUpClusterDescriptor(ImmutableList.of(), new AbortException(), null);
-    FormValidation result =
-        descriptor.doCheckCluster(
-            jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_CredentialAuthFailed(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithIOException() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, new IOException());
-    FormValidation result =
-        descriptor.doCheckCluster(
-            jenkins, TEST_CLUSTER, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_ClusterVerificationError(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithAbortExceptionAndEmptyCluster() throws IOException {
-    DescriptorImpl descriptor =
-        setUpClusterDescriptor(ImmutableList.of(), new AbortException(), null);
-    FormValidation result =
-        descriptor.doCheckCluster(jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_CredentialAuthFailed(), result.getMessage());
-  }
-
-  @Test
-  public void testDoCheckClusterMessageWithIOExceptionAndEmptyCluster() throws IOException {
-    DescriptorImpl descriptor = setUpClusterDescriptor(ImmutableList.of(), null, new IOException());
-    FormValidation result =
-        descriptor.doCheckCluster(jenkins, null, TEST_CREDENTIALS_ID, TEST_PROJECT_ID, EMPTY_VALUE);
-    assertNotNull(result);
-    assertEquals(Messages.KubernetesEngineBuilder_ClusterVerificationError(), result.getMessage());
-  }
-
-  @Test
   public void testDoCheckVerifyTimeoutInMinutesNAN() {
     DescriptorImpl descriptor = Mockito.spy(DescriptorImpl.class);
     FormValidation result = descriptor.doCheckVerifyTimeoutInMinutes("abc");
@@ -803,40 +532,7 @@ public class KubernetesEngineBuilderTest {
     return descriptor;
   }
 
-  private DescriptorImpl setUpClusterDescriptor(
-      List<String> initialClusters, AbortException abortException, IOException ioException)
-      throws IOException {
-    DescriptorImpl descriptor = Mockito.spy(DescriptorImpl.class);
-
-    if (abortException != null) {
-      Mockito.doThrow(abortException)
-          .when(descriptor)
-          .getClientFactory(any(Jenkins.class), anyString());
-      return descriptor;
-    }
-
-    ClientFactory clientFactory = Mockito.mock(ClientFactory.class);
-    Mockito.doReturn(clientFactory)
-        .when(descriptor)
-        .getClientFactory(any(Jenkins.class), anyString());
-
-    ContainerClient containerClient = Mockito.mock(ContainerClient.class);
-    Mockito.when(clientFactory.containerClient()).thenReturn(containerClient);
-
-    if (ioException != null) {
-      Mockito.when(containerClient.listClusters(anyString(), anyString())).thenThrow(ioException);
-      return descriptor;
-    }
-
-    List<Cluster> clusters = new ArrayList<>();
-    initialClusters.forEach(c -> clusters.add(ContainerClient.fromNameAndZone(c)));
-    Mockito.when(containerClient.listClusters(anyString(), anyString()))
-        .thenReturn(ImmutableList.copyOf(clusters));
-    return descriptor;
-  }
-
-  private static ListBoxModel initExpected(
-      List<String> expectedNames, List<String> expectedValues) {
+  static ListBoxModel initExpected(List<String> expectedNames, List<String> expectedValues) {
     ListBoxModel expected = new ListBoxModel();
     for (int i = 0; i < expectedNames.size(); i++) {
       expected.add(expectedNames.get(i), expectedValues.get(i));
@@ -844,14 +540,14 @@ public class KubernetesEngineBuilderTest {
     return expected;
   }
 
-  private static void assertListBoxModelEquals(ListBoxModel expected, ListBoxModel result) {
+  static void assertListBoxModelEquals(ListBoxModel expected, ListBoxModel result) {
     for (int i = 0; i < expected.size(); i++) {
       assertEquals(expected.get(i).name, result.get(i).name);
       assertEquals(expected.get(i).value, result.get(i).value);
     }
   }
 
-  private static void assertValueSelected(ListBoxModel result, String expectedValue) {
+  static void assertValueSelected(ListBoxModel result, String expectedValue) {
     Optional<Option> expectedOption =
         result.stream().filter(i -> expectedValue.equals(i.value)).findFirst();
     expectedOption.ifPresent(option -> assertTrue(option.selected));
